Rendering something on the swapchain.

+++ source/shaders/global_ubo.glsl
@@ -4,6 +4,6 @@ layout(set = 0, binding = 0) uniform GlobalUbo {
   mat4 view_proj;
   vec4 view_dir;
 
-  ivec2 view_dim;
+  vec2 view_dim;
 }
 global_ubo;
